requirements.txt
amqp==5.3.1
asgiref==3.8.0
billiard==4.2.1
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.3
django-cors-headers==4.3.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
idna==3.6
kombu==5.4.2
oauthlib==3.2.2
pillow==10.2.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
redis==5.2.1
requests==2.31.0
requests-oauthlib==1.4.0
six==1.17.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
stripe==8.8.0
typing_extensions==4.10.0
tzdata==2024.2
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
